requirements.txt
bit>=0.4.3
django>=2.0.3
django-countries>=5.1.1
web3>=4.1.0
psycopg2>=2.7.4
psycopg2-binary>=2.7.4
django-filter>=1.1.0
django-extensions>=2.0.6
ethereum>=2.3.1
django-otp>=0.4.3
django-registration-redux>=2.2
django-two-factor-auth>=1.7.0
django_chartit>=0.2.9
ecdsa>=0.13
celery>=4.1.0
django-qrcode>=0.3
eth-utils>=0.8.1
django-fernet-fields>=0.5
redis>=2.10.6
rlp==0.6.0
coinbase>=2.1.0
currencyconverter>=0.13.3
